+++ packages/saas-ui-forms/zod/src/create-zod-form.ts
@@ -3,14 +3,17 @@ import {
   CreateFormProps,
   WithFields,
   FormProps,
+  GetBaseField,
 } from '@saas-ui/forms'
 import { zodFieldResolver, zodResolver } from './zod-resolver'
 import { z } from 'zod'
 
 type ResolverArgs = Parameters<typeof zodResolver>
 
-export interface CreateZodFormProps<FieldDefs>
-  extends CreateFormProps<FieldDefs> {
+export interface CreateZodFormProps<
+  FieldDefs,
+  TGetBaseField extends GetBaseField = GetBaseField,
+> extends CreateFormProps<FieldDefs, TGetBaseField> {
   schemaOptions?: ResolverArgs[1]
   resolverOptions?: ResolverArgs[2]
 }
@@ -18,15 +21,16 @@ export interface CreateZodFormProps<FieldDefs>
 export type ZodFormType<
   FieldDefs,
   ExtraProps = object,
+  ExtraFieldProps extends object = object,
   ExtraOverrides = object,
-  Type extends 'zod' = 'zod'
+  Type extends 'zod' = 'zod',
 > = (<
   TSchema extends z.AnyZodObject = z.AnyZodObject,
   TFieldValues extends z.infer<TSchema> = z.infer<TSchema>,
-  TContext extends object = object
+  TContext extends object = object,
 >(
   props: WithFields<
-    FormProps<TSchema, TFieldValues, TContext>,
+    FormProps<TSchema, TFieldValues, TContext, ExtraFieldProps>,
     FieldDefs,
     ExtraOverrides
   > & {
@@ -37,9 +41,17 @@ export type ZodFormType<
   id?: string
 }
 
-export const createZodForm = <FieldDefs>(
-  options?: CreateZodFormProps<FieldDefs>
+export const createZodForm = <
+  FieldDefs,
+  TGetBaseField extends GetBaseField<any> = GetBaseField<any>,
+>(
+  options?: CreateZodFormProps<FieldDefs, TGetBaseField>
 ) => {
+  type ExtraFieldProps =
+    TGetBaseField extends GetBaseField<infer ExtraFieldProps>
+      ? ExtraFieldProps
+      : object
+
   const ZodForm = createForm({
     resolver: (schema: any) =>
       zodResolver(schema, options?.schemaOptions, options?.resolverOptions),
@@ -50,5 +62,5 @@ export const createZodForm = <FieldDefs>(
   ZodForm.displayName = 'ZodForm'
   ZodForm.id = 'ZodForm'
 
-  return ZodForm as ZodFormType<FieldDefs>
+  return ZodForm as ZodFormType<FieldDefs, object, ExtraFieldProps>
 }