+++ apps/website/CHANGELOG.md
@@ -1,5 +1,20 @@
 # website
 
+## 2.3.25
+
+### Patch Changes
+
+- @saas-ui/forms@2.5.1
+- @saas-ui/react@2.7.1
+
+## 2.3.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/forms@2.5.0
+  - @saas-ui/react@2.7.0
+
 ## 2.3.23
 
 ### Patch Changes