+++ packages/saas-ui-modals/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @saas-ui/modals
 
+## 2.1.26
+
+### Patch Changes
+
+- @saas-ui/forms@2.5.1
+
+## 2.1.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/forms@2.5.0
+
 ## 2.1.24
 
 ### Patch Changes