+++ packages/saas-ui-react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @saas-ui/react
 
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/core@2.4.2
+  - @saas-ui/data-table@9.0.1
+  - @saas-ui/forms@2.5.1
+  - @saas-ui/hotkeys@2.2.5
+  - @saas-ui/modals@2.1.26
+
+## 2.7.0
+
+### Minor Changes
+
+- Improved forms to support a global base field override
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/forms@2.5.0
+  - @saas-ui/modals@2.1.25
+  - @saas-ui/data-table@9.0.0
+
 ## 2.6.2
 
 ### Patch Changes