+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,18 @@
 # saas-ui-storybook
 
+## 1.1.27
+
+### Patch Changes
+
+- @saas-ui/react@2.7.1
+
+## 1.1.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/react@2.7.0
+
 ## 1.1.25
 
 ### Patch Changes