+++ tooling/test-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @saas-ui/test-utils
 
+## 3.1.27
+
+### Patch Changes
+
+- @saas-ui/react@2.7.1
+
+## 3.1.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/react@2.7.0
+
 ## 3.1.25
 
 ### Patch Changes